Whew, updated to make it easily generalizable so...
that I can run Hopper and Walker on it. Then this will be the basis for the TRPO code. Also, I can't figure out why the KL divergence is sometimes negative, but is that just a discretization error? I did double check the KL divergence (there was a mistake I had earlier with -d, whereas we really just need -1 due to the reduce sum across the rows, very clever) and I checked the entropy and the gaussian pdf, looks OK and generalizes to more than one action dimension, but not to non-diagonal covariance matrices.